+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -120,6 +120,21 @@ export async function loadMetroConfigAsync(
   const serverRoot = getMetroServerRoot(projectRoot);
   const terminalReporter = new MetroTerminalReporter(serverRoot, terminal);
 
+  const hasConfig = await resolveConfig(options.config, projectRoot);
+  let config: ConfigT = await loadConfig(
+    { cwd: projectRoot, projectRoot, ...options },
+    // If the project does not have a metro.config.js, then we use the default config.
+    hasConfig.isEmpty ? getDefaultConfig(projectRoot) : undefined
+  )
+
+  const configReporter = config.reporter;
+  config = {
+    ...config,
+    reporter : {
+      update(event: any) {
+        if(configReporter){
+          configReporter.update(event);
+        }
   const defaultConfig = getDefaultConfig(projectRoot);
   const resolvedConfig = await resolveConfig(options.config, projectRoot);
 
@@ -150,8 +165,8 @@ export async function loadMetroConfigAsync(
         if (reportEvent) {
           reportEvent(event);
         }
-      },
-    },
+      }
+    }
   };
 
   globalThis.__requireCycleIgnorePatterns = config.resolver?.requireCycleIgnorePatterns;